+++ changelog.md
@@ -81,3 +81,6 @@
 
 ### *v.023* (April 1, 2017)
 - Add API key function (Jake)
+
+### *v.024* (April 1, 2017)
+- Added the reboot function (Karan)